New exportBook: no clashing identifiers, real metadata, and true xml parsing
Previously, if a chapter had a similar filename to a section or an
image, they would be given the same identifier in the bookizip, which
had strange effects in objavi.

The previous metadata was entirely fake and the same for all books.
Now it is based on actually recorded metadata.

Elements with src attributes (eg images) that link to local files
often need those attribute values to be rewritten for the bookizip.
This used to be done with regular expressions, but now the xml is
parsed.  This also affects the way encoded unicode characters are
treated.

The code has also been reorganised and split into several functions.

This work was done over several commits in the bookzip branch, but
there is no easy way to cherrypick those, due to significant changes
in common.py in the rome branch.