Add list-books-by-id/<scheme>.json for Internet Archive interface
The internet archive needs to be able to query books by their
archive IDs.  To do this we make an intermediate server that
periodically asks booki for a list of books.  This patch adds
the list of books to Booki.

To get a list of books which have an ID in the 'archive.org'
scheme, visit an url like this:

http://$BOOKISERVER/list-books-by-id/archive.org.json

This will give you a JSON file in this format:

{
  "archive-id-of-an-unchanged-book" : ["url to edit", null],
  "archive-id-of-a-changed-book"    : ["url to edit", "url to download"],
}